+++ ResellKH/ResellKH/src/main/java/com/example/resellkh/controller/NotificationController.java
@@ -1,7 +1,6 @@
 package com.example.resellkh.controller;
 
 import com.example.resellkh.model.dto.ApiResponse;
-import com.example.resellkh.model.dto.NotificationRequest;
 import com.example.resellkh.model.entity.Notification;
 import com.example.resellkh.service.Impl.NotificationServiceImpl;
 import lombok.RequiredArgsConstructor;
@@ -20,7 +19,7 @@ public class NotificationController {
     private final NotificationServiceImpl notificationServiceImpl;
 
     @GetMapping("/{userId}")
-    public ResponseEntity<ApiResponse<Notification>> getNotificationByUserId(@PathVariable int userId) {
+    public ResponseEntity<ApiResponse<Notification>> getNotificationByUserId(@PathVariable Long userId) {
         Notification notification = notificationServiceImpl.getNotificationByUserId(userId);
 
         if (notification == null) {
@@ -35,7 +34,7 @@ public class NotificationController {
     }
 
     @GetMapping("/all/{userId}")
-    public ResponseEntity<ApiResponse<Iterable<Notification>>> getAllNotificationsByUserId(@PathVariable int userId) {
+    public ResponseEntity<ApiResponse<Iterable<Notification>>> getAllNotificationsByUserId(@PathVariable Long userId) {
         List<Notification> notifications = notificationServiceImpl.getAllNotificationsByUserId(userId);
         if (notifications == null || notifications.isEmpty()) {
             return ResponseEntity.ok(
@@ -48,14 +47,14 @@ public class NotificationController {
     }
 
     @PutMapping("/read/{userId}/{id}")
-    public ResponseEntity<String> markNotificationsAsRead(@PathVariable int userId, @PathVariable int id) {
+    public ResponseEntity<String> markNotificationsAsRead(@PathVariable Long userId, @PathVariable Long id) {
         notificationServiceImpl.markNotificationAsRead(userId, id);
         return ResponseEntity.ok("Notifications marked as read.");
     }
 
     @GetMapping("/getproductidbynotid/{id}")
-    public ResponseEntity<ApiResponse<Integer>> getProductId(@PathVariable int id) {
-        Integer productId = notificationServiceImpl.getProductIdByNoId(id);
+    public ResponseEntity<ApiResponse<Long>> getProductId(@PathVariable Long id) {
+        Long productId = notificationServiceImpl.getProductIdByNoId(id);
 
         if (productId == null) {
             // If service returns null, it means not found. Return 404.
