+++ ResellKH/ResellKH/src/main/java/com/example/resellkh/model/entity/Notification.java
@@ -12,9 +12,9 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @Builder
 public class Notification {
-    private int id;
-    private int userId;
-    private Integer productId; // Changed from int to Integer
+    private Long id;
+    private Long userId;
+    private Long productId; // Changed from int to Integer
     private String title;
     private String content;
     private String iconUrl;
